Foundry_Rag/rag_embedding/requirements.txt
azure-ai-projects>=2.0.0
azure-identity
python-dotenv
openai 
num2words 
matplotlib 
plotly 
scipy 
scikit-learn 
pandas 
tiktoken